backend/requirements.txt
fastapi==0.104.1
pydantic==2.4.2
uvicorn==0.24.0
bcrypt==4.0.1
cryptography==41.0.5
passlib==1.7.4
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
requests==2.26.0